requirements.txt
flask
werkzeug
boto3
pillow
gunicorn
requests
pyyaml